requirements.txt
fastapi
boto3
PyPDF2
streamlit
google-generativeai
fastapi
pymongo
python-dotenv
PyMuPDF
uvicorn
requests
pinecone
pdf2image